rename sub package
