+++ tests/apps/events/sync_database/sync_database_test.py
@@ -1,6 +1,6 @@
 import pytest
 
-from busy_beaver.apps.events.event_database.sync_database import (
+from busy_beaver.apps.events.sync_database.sync_database import (
     classify_transaction_type,
 )
